feat: Add deletePlayers function to delete multiple players

+++ app/player/columns.tsx
@@ -4,8 +4,19 @@ import { Checkbox } from "@/components/ui/Checkbox";
 import LinkButton from "@/components/LinkButton";
 import { Button } from "@/components/ui/Button";
 import type { ColumnDef } from "@tanstack/react-table";
-import { toast } from "sonner";
 import { ArrowUpDown } from "lucide-react";
+import { AlertDialogHeader, AlertDialogFooter } from "@/components/ui/Alert-dialog";
+
+import {
+  AlertDialog,
+  AlertDialogTrigger,
+  AlertDialogContent,
+  AlertDialogTitle,
+  AlertDialogDescription,
+  AlertDialogCancel,
+  AlertDialogAction,
+} from "@radix-ui/react-alert-dialog";
+import { toast } from "sonner";
 import { deletePlayer } from "./actions";
 
 export type Team = {
@@ -74,21 +85,31 @@ export const columns: ColumnDef<Players>[] = [
     accessorKey: "delete",
     header: "Delete",
     cell: ({ row }) => (
-      <Button
-        onClick={() => {
-          toast("are you sure you want to delete?", {
-            action: {
-              label: "Delete",
-              onClick: async () => {
+      <AlertDialog>
+        <AlertDialogTrigger asChild>
+          <Button>delete</Button>
+        </AlertDialogTrigger>
+        <AlertDialogContent>
+          <AlertDialogHeader>
+            <AlertDialogTitle>Are you absolutely sure?</AlertDialogTitle>
+            <AlertDialogDescription>
+              This action cannot be undone. This will permanently delete this player and remove their data
+            </AlertDialogDescription>
+          </AlertDialogHeader>
+          <AlertDialogFooter>
+            <AlertDialogCancel>Cancel</AlertDialogCancel>
+            <AlertDialogAction
+              onClick={async () => {
                 await deletePlayer(row.original.id);
-                window.location.reload(); // trouver une autre solution
-              },
-            },
-          });
-        }}
-      >
-        delete
-      </Button>
+                toast.success("player deleted successfully");
+                window.location.reload();
+              }}
+            >
+              Delete
+            </AlertDialogAction>
+          </AlertDialogFooter>
+        </AlertDialogContent>
+      </AlertDialog>
     ),
   },
 ];
